requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
pydantic>=2.5.0
google-cloud-firestore>=2.13.0
google-cloud-storage>=2.14.0
google-cloud-secret-manager>=2.16.0
google-cloud-pubsub>=2.21.0
google-auth>=2.28.0
google-api-python-client>=2.120.0
weasyprint>=61.0
python-dotenv>=1.0.0
